realizar publicaciones con imagenes, mejorar estilos del perfil
